requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
groq
tavily-python
